# contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ComposeFile": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "$ref": "#/components/schemas/ComposeFileName"
          },
          "sha256": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "sha256"
        ],
        "type": "object"
      },
...
      "ComposeImpact": {
        "additionalProperties": false,
        "properties": {
          "elevated": {
            "items": {
              "$ref": "#/components/schemas/ComposePrivilege"
            },
            "type": "array"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ComposeServiceChange"
            },
            "type": "array"
          }
        },
        "required": [
          "services",
          "elevated"
        ],
        "type": "object"
      },
...
      "ComposePlan": {
        "additionalProperties": false,
        "properties": {
          "before": {
            "$ref": "#/components/schemas/ComposeProject"
          },
          "catalog_digest": {
            "type": "string"
          },
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "desired": {
            "$ref": "#/components/schemas/ComposeProject"
          },
          "expires_at": {
            "format": "int64",
            "type": "integer"
          },
          "grant_revision": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "impact": {
            "$ref": "#/components/schemas/ComposeImpact"
          },
          "managed_file": {
            "type": "string"
          },
          "operator_files": {
            "items": {
              "$ref": "#/components/schemas/ComposeFile"
            },
            "type": "array"
          },
          "principal": {
            "type": "string"
          },
          "selection": {
            "$ref": "#/components/schemas/ComposeSelection"
          },
          "template_digest": {
            "type": "string"
          },
          "version": {
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "principal",
          "grant_revision",
          "catalog_digest",
          "template_digest",
          "selection",
          "operator_files",
          "managed_file",
          "before",
          "desired",
          "impact",
          "created_at",
          "expires_at"
        ],
        "type": "object"
      },
...
      "ComposePrivilege": {
        "enum": [
          "privileged",
          "host_mount",
          "device",
          "docker_socket",
          "host_network"
        ],
        "type": "string"
      },
      "ComposeProject": {
        "additionalProperties": false,
        "properties": {
          "services": {
            "items": {
              "$ref": "#/components/schemas/ComposeService"
            },
            "type": "array"
          }
        },
        "required": [
          "services"
        ],
        "type": "object"
      },
...
      "ComposeSelection": {
        "additionalProperties": false,
        "properties": {
          "stack": {
            "type": "string"
          },
          "template": {
            "type": "string"
          }
        },
        "required": [
          "stack",
          "template"
        ],
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "invalid_input",
          "unauthenticated",
          "forbidden",
          "conflict",
          "expired",
          "overloaded",
          "unavailable",
          "corrupt_configuration",
          "state_not_durable",
          "not_found"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "session": {
        "in": "cookie",
        "name": "__Host-limeos",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "LimeOS foundation API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/compose/plans": {
      "post": {
        "description": "Select from the protected catalog and persist a non-executable preview. Deployment requires a new P04 operation and fresh approval.",
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 64,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "additionalProperties": false,
                "properties": {
                  "stack": {
                    "type": "string"
                  },
                  "template": {
                    "type": "string"
                  }
                },
                "required": [
                  "stack",
                  "template"
                ],
                "title": "ComposeSelection",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "digest": {
                      "type": "string"
                    },
                    "plan": {
                      "$ref": "#/components/schemas/ComposePlan"
                    }
                  },
                  "required": [
                    "plan",
                    "digest"
                  ],
                  "title": "PlannedCompose",
                  "type": "object"
                }
              }
            },
            "description": "Durable plan, approval or queued job"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/compose/plans/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[a-f0-9]{64}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "digest": {
                      "type": "string"
                    },
                    "plan": {
                      "$ref": "#/components/schemas/ComposePlan"
                    }
                  },
                  "required": [
                    "plan",
                    "digest"
                  ],
                  "title": "PlannedCompose",
                  "type": "object"
                }
              }
            },
            "description": "Current owner-scoped preview; no host effects or expiry refresh"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/compose/plans/{id}/approval": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 64,
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^[a-f0-9]{64}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "additionalProperties": false,
                "properties": {
                  "digest": {
                    "type": "string"
                  }
                },
                "required": [
                  "digest"
                ],
                "title": "ApprovalInput",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "expires_at": {
                      "format": "int64",
                      "type": "integer"
                    },
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token",
                    "expires_at"
                  ],
                  "title": "PlanApproval",
                  "type": "object"
                }
              }
            },
            "description": "Durable plan, approval or queued job"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "audit_id": {
                      "type": "string"
                    },
                    "code": {
                      "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retry",
                    "audit_id"
                  ],
                  "title": "ErrorEnvelope",
                  "type": "object"
                }
              }
            },
            "description": "Structured safe error"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  }
}